% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepfmri.R
\name{prepfmri}
\alias{prepfmri}
\title{preprocessing fMRI data}
\usage{
prepfmri(nii, mask, r = 40)
}
\arguments{
\item{nii}{A 4-D fMRI image.}

\item{mask}{A 3-D mask has the same dimension with \code{dim(nii)[1:3]}.}

\item{r}{The number of PCs.}
}
\value{
The preprocessed n*r data matrix (n is the number of masked voxels).
}
\description{
The function preprocesses fMRI data (NIfTI format) with the following steps:
detrending; demeaning; and PCA.
}
