% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ics.R
\name{write_ics}
\alias{write_ics}
\title{Write the estimated components into a NIfTI file.}
\usage{
write_ics(ICs, nii, mask, q = 0.99, file)
}
\arguments{
\item{ICs}{The output of DICA. A}

\item{nii}{The reference image.}

\item{mask}{A 3-D mask.}

\item{q}{The quantile thresholding the components.}
}
\value{

}
\description{
Write the estimated components into a NIfTI file.
}
