% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dica_2stage.R
\name{dica}
\alias{dica}
\title{Distributional ICA}
\usage{
dica(
  Y,
  K = 6,
  L = 3,
  Rmat = diag(L),
  tol = 1e-05,
  itr.max = 1000,
  log = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Y}{A numeric matrix or a list. If a matrix, a mixture of Gaussian model
is to be fitted at stage one, and rows correspond to observations (\code{n}) and columns
correspond to the dimension (\code{d}). If a list, a mixture of Wishart model
is to be fitted at stage one. Each item of the list is a \code{d*d} positive definite
matrix, corresponding to a observation.}

\item{K}{The number of components at stage one.}

\item{L}{The number of components at stage two, should be no larger than \code{K}.}

\item{Rmat}{Initial estimate of the \code{L*L} orthogonal rotation matrix at stage two.}

\item{tol}{Relative convergence tolerance for the log-likelihood.}

\item{itr.max}{Maximum number of algorithm iterations.}

\item{log}{if \code{TRUE}, the posterior probabilities at stage one are calculated on the log scale.}

\item{verbose}{If \code{TRUE}, print progress of algorithm to console.}
}
\value{
A list includes
\itemize{
\item \code{S} - The estimated component matrix.
\item \code{A} - The estiamted mixing matrix.
}
}
\description{
Distributional independent component ICA with two-stage estimation.
At stage one, a mixture model (Gaussian or Wishart) is estimated by EM algorithm. At stage two, the posterior
weights (after a transformation with the multinomial logit link function)
are decomposed with Infomax ICA.
}
