% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dti.R
\name{get_dti}
\alias{get_dti}
\title{Reconstrct DTI observations.}
\usage{
get_dti(V1, V2, V3, L1, L2, L3, mask)
}
\arguments{
\item{mask}{A 3-D mask has the same dimension with the image data.}
}
\value{

}
\description{
Reconstrct DTI observations.
}
